Moving contents out of tensorflow_models
