+++ contrib/vina_model/1jld_protein.pdb
