+++ deepchem/models/__init__.py
@@ -10,11 +10,11 @@ from deepchem.models.sklearn_models import SklearnModel
 from deepchem.models.xgboost_models import XGBoostModel
 from deepchem.models.multitask import SingletaskToMultitask
 
-from deepchem.models.tensorflow_models.fcnet import MultiTaskRegressor
-from deepchem.models.tensorflow_models.fcnet import MultiTaskClassifier
-from deepchem.models.tensorflow_models.fcnet import MultiTaskFitTransformRegressor
-from deepchem.models.tensorflow_models.IRV import TensorflowMultiTaskIRVClassifier
 from deepchem.models.tensorgraph.tensor_graph import TensorGraph
+from deepchem.models.tensorgraph.fcnet import MultiTaskRegressor
+from deepchem.models.tensorgraph.fcnet import MultiTaskClassifier
+from deepchem.models.tensorgraph.fcnet import MultiTaskFitTransformRegressor
+from deepchem.models.tensorgraph.IRV import TensorflowMultiTaskIRVClassifier
 from deepchem.models.tensorgraph.models.graph_models import WeaveTensorGraph, DTNNTensorGraph, DAGTensorGraph, GraphConvTensorGraph, MPNNTensorGraph
 from deepchem.models.tensorgraph.models.symmetry_function_regression import BPSymmetryFunctionRegression, ANIRegression
