+++ deepchem/models/tensorflow_models/progressive_multitask.py
@@ -1,3 +1,4 @@
+'''
 from __future__ import print_function
 from __future__ import division
 from __future__ import unicode_literals
@@ -11,7 +12,6 @@ from deepchem.utils.save import log
 from deepchem.metrics import to_one_hot
 from deepchem.metrics import from_one_hot
 from deepchem.nn import model_ops
-'''
 class ProgressiveMultitaskRegressor(TensorflowMultiTaskRegressor):
   """Implements a progressive multitask neural network.
