+++ deepchem/models/tensorflow_models/robust_multitask.py
@@ -1,3 +1,4 @@
+'''
 from __future__ import print_function
 from __future__ import division
 from __future__ import unicode_literals
@@ -7,7 +8,6 @@ import numpy as np
 import tensorflow as tf
 
 from deepchem.nn import model_ops
-'''
 class RobustMultitaskClassifier(MultiTaskClassifier):
   """Implements a neural network for robust multitasking.
